+++ HISTORY.rst
@@ -7,6 +7,8 @@ Unreleased
 
 * BUGFIX: Provides a helpful error message when `geckoview_datapoint` is used on an metric type that doesn't support GeckoView exfiltration.
 
+* Generate a lookup table for Gecko categorical histograms in `GleanGeckoMetricsMapping`.
+
 1.4.1 (2019-08-28)
 ------------------